+++ entities/article/article.schema.js
