+++ routes/index.js
@@ -1,5 +1,5 @@
 import express from "express";
-import { currentUser } from "../user/user.controller";
+import { currentUser } from "../entities/user/user.controller.js";
 const router = express.Router();
 
 router.get("/users/me", currentUser);
